+++ app/components/primer/underline_nav_component.rb
@@ -36,7 +36,7 @@ module Primer
     #
     # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
     renders_one :actions, lambda { |**system_arguments|
-      system_arguments[:tag] ||= :div
+      system_arguments[:tag] ||= :div # rubocop:disable Primer/NoTagMemoize
       system_arguments[:classes] = class_names("UnderlineNav-actions", system_arguments[:classes])
 
       Primer::BaseComponent.new(**system_arguments)